+++ lib/commanded/aggregates/aggregate.ex
@@ -99,18 +99,18 @@ defmodule Commanded.Aggregates.Aggregate do
 
   require Logger
 
+  alias Commanded.Aggregate.Multi
   alias Commanded.Aggregates.Aggregate
+  alias Commanded.Aggregates.AggregateStateBuilder
   alias Commanded.Aggregates.ExecutionContext
+  alias Commanded.Application.Config
   alias Commanded.Event.Mapper
   alias Commanded.Event.Upcast
   alias Commanded.EventStore
   alias Commanded.EventStore.RecordedEvent
-  alias Commanded.EventStore.SnapshotData
   alias Commanded.Registration
   alias Commanded.Snapshotting
   alias Commanded.Telemetry
-
-  @read_event_batch_size 100
 
   defstruct [
     :application,
@@ -197,7 +197,38 @@ defmodule Commanded.Aggregates.Aggregate do
   @doc false
   def aggregate_state(application, aggregate_module, aggregate_uuid, timeout \\ 5_000) do
     name = via_name(application, aggregate_module, aggregate_uuid)
-    GenServer.call(name, :aggregate_state, timeout)
+
+    try do
+      GenServer.call(name, :aggregate_state, timeout)
+    catch
+      :exit, {reason, {GenServer, :call, [^name, :aggregate_state, ^timeout]}}
+      when reason in [:normal, :noproc] ->
+        task =
+          Task.async(fn ->
+            snapshot_options =
+              application
+              |> Config.get(:snapshotting)
+              |> Kernel.||(%{})
+              |> Map.get(aggregate_module, [])
+
+            %Aggregate{
+              application: application,
+              aggregate_module: aggregate_module,
+              aggregate_uuid: aggregate_uuid,
+              snapshotting: Snapshotting.new(application, aggregate_uuid, snapshot_options)
+            }
+            |> AggregateStateBuilder.populate()
+            |> Map.fetch!(:aggregate_state)
+          end)
+
+        case Task.yield(task, timeout) || Task.shutdown(task) do
+          {:ok, result} ->
+            result
+
+          nil ->
+            exit({:timeout, {GenServer, :call, [name, :aggregate_state, timeout]}})
+        end
+    end
   end
 
   @doc false
@@ -229,9 +260,11 @@ defmodule Commanded.Aggregates.Aggregate do
   @doc false
   @impl GenServer
   def handle_continue(:populate_aggregate_state, %Aggregate{} = state) do
+    state = AggregateStateBuilder.populate(state)
+
     # Subscribe to aggregate's events to catch any events appended to its stream
     # by another process, such as directly appended to the event store.
-    {:noreply, populate_aggregate_state(state), {:continue, :subscribe_to_events}}
+    {:noreply, state, {:continue, :subscribe_to_events}}
   end
 
   @doc false
@@ -410,94 +443,6 @@ defmodule Commanded.Aggregates.Aggregate do
     end
   end
 
-  # Populate the aggregate's state from a snapshot, if present, and it's events.
-  #
-  # Attempt to fetch a snapshot for the aggregate to use as its initial state.
-  # If the snapshot exists, fetch any subsequent events to rebuild its state.
-  # Otherwise start with the aggregate struct and stream all existing events for
-  # the aggregate from the event store to rebuild its state from those events.
-  defp populate_aggregate_state(%Aggregate{} = state) do
-    %Aggregate{aggregate_module: aggregate_module, snapshotting: snapshotting} = state
-
-    aggregate =
-      case Snapshotting.read_snapshot(snapshotting) do
-        {:ok, %SnapshotData{source_version: source_version, data: data}} ->
-          %Aggregate{
-            state
-            | aggregate_version: source_version,
-              aggregate_state: data
-          }
-
-        {:error, _error} ->
-          # No snapshot present, or exists but for outdated state, so use intial empty state
-          %Aggregate{state | aggregate_version: 0, aggregate_state: struct(aggregate_module)}
-      end
-
-    rebuild_from_events(aggregate)
-  end
-
-  # Load events from the event store, in batches, to rebuild the aggregate state
-  defp rebuild_from_events(%Aggregate{} = state) do
-    %Aggregate{
-      application: application,
-      aggregate_uuid: aggregate_uuid,
-      aggregate_version: aggregate_version
-    } = state
-
-    case EventStore.stream_forward(
-           application,
-           aggregate_uuid,
-           aggregate_version + 1,
-           @read_event_batch_size
-         ) do
-      {:error, :stream_not_found} ->
-        # aggregate does not exist, return initial state
-        state
-
-      event_stream ->
-        rebuild_from_event_stream(event_stream, state)
-    end
-  end
-
-  # Rebuild aggregate state from a `Stream` of its events.
-  defp rebuild_from_event_stream(event_stream, %Aggregate{} = state) do
-    Enum.reduce(event_stream, state, fn event, state ->
-      %RecordedEvent{data: data, stream_version: stream_version} = event
-
-      %Aggregate{
-        aggregate_module: aggregate_module,
-        aggregate_state: aggregate_state,
-        aggregate_version: aggregate_version,
-        snapshotting: snapshotting
-      } = state
-
-      state_with_snapshot =
-        if snapshotting && Snapshotting.snapshot_required?(snapshotting, stream_version) do
-          case Snapshotting.take_snapshot(snapshotting, aggregate_version, aggregate_state) do
-            {:ok, snapshotting} ->
-              # nocommit
-              IO.puts("Stream version: #{stream_version}")
-              %Aggregate{state | snapshotting: snapshotting}
-
-            {:error, error} ->
-              Logger.warn(fn ->
-                describe(state) <> " snapshot failed due to: " <> inspect(error)
-              end)
-
-              state
-          end
-        else
-          state
-        end
-
-      %Aggregate{
-        state_with_snapshot
-        | aggregate_version: stream_version,
-          aggregate_state: aggregate_module.apply(aggregate_state, data)
-      }
-    end)
-  end
-
   defp aggregate_lifespan_timeout(lifespan, function_name, args) do
     # Take the last event or the command or error
     args = args |> List.wrap() |> Enum.take(-1)
@@ -549,8 +494,8 @@ defmodule Commanded.Aggregates.Aggregate do
         none when none in [:ok, nil, []] ->
           {{:ok, []}, state}
 
-        %Commanded.Aggregate.Multi{} = multi ->
-          case Commanded.Aggregate.Multi.run(multi) do
+        %Multi{} = multi ->
+          case Multi.run(multi) do
             {:error, _error} = reply ->
               {reply, state}
 
@@ -605,7 +550,7 @@ defmodule Commanded.Aggregates.Aggregate do
     else
       {:error, :wrong_expected_version} ->
         # Fetch missing events from event store
-        state = rebuild_from_events(state)
+        state = AggregateStateBuilder.rebuild_from_events(state)
 
         # Retry command if there are any attempts left
         case ExecutionContext.retry(context) do