+++ test/event_store/support/subscription_test_case.ex
@@ -2,9 +2,7 @@ defmodule Commanded.EventStore.SubscriptionTestCase do
   import Commanded.SharedTestCase
 
   define_tests do
-    alias Commanded.EventStore.EventData
-    alias Commanded.EventStore.Subscriber
-    alias Commanded.EventStore.RecordedEvent
+    alias Commanded.EventStore.{EventData, RecordedEvent, Subscriber}
     alias Commanded.Helpers.ProcessHelper
 
     defmodule BankAccountOpened do
@@ -187,7 +185,7 @@ defmodule Commanded.EventStore.SubscriptionTestCase do
         refute_receive {:events, _events}
       end
 
-      test "should receive events already apended to stream", %{
+      test "should receive events already appended to stream", %{
         event_store: event_store,
         event_store_meta: event_store_meta
       } do
@@ -530,7 +528,9 @@ defmodule Commanded.EventStore.SubscriptionTestCase do
           :ok
 
         subscription ->
-          event_store.ack_event(event_store_meta, subscription, List.last(received_events))
+          last_event = List.last(received_events)
+
+          event_store.ack_event(event_store_meta, subscription, last_event)
       end
 
       case expected_count - length(received_events) do